editClass
