+++ src/components/Class/ClassCard/index.js
@@ -12,37 +12,20 @@ import "./styles.scss";
 import { toast } from "react-toastify";
 import "react-toastify/dist/ReactToastify.css";
 import Loader from "../../common/Loader";
-import FormControlLabel from "@mui/material/FormControlLabel";
-import {
-  Typography,
-  Card,
-  Grid,
-  Button,
-  Box,
-  Stack,
-  Menu,
-  MenuItem,
-  Checkbox,
-  CardActions,
-  Dialog,
-  DialogTitle,
-  DialogActions,
-} from "@mui/material";
+import { Typography, Card, Grid, Button, CardActions } from "@mui/material";
 import useMediaQuery from "@mui/material/useMediaQuery";
 
-import MoreVertIcon from "@mui/icons-material/MoreVert";
-import ExternalLink from "../../common/ExternalLink";
 import ClassJoinTimerButton from "../ClassJoinTimerButton";
 import MergeClass from "../MergeClass";
 import ArrowRightAltIcon from "@mui/icons-material/ArrowRightAlt";
 import { useParams } from "react-router-dom";
+import EditClass from "../EditClass";
 toast.configure();
 
 function ClassCard() {
   const params = useParams();
   console.log(params);
   const classId = params.batchId;
-  const PathwayID = params.pathwayID;
 
   const classes = useStyles();
   const dispatch = useDispatch();
@@ -83,48 +66,9 @@ function ClassCard() {
     doubt_class: "Doubt Class",
   };
 
-  const handleOpenUserMenu = (event) => {
-    setAnchorElUser(event.currentTarget);
-  };
-
-  const handleClose = () => {
-    setShowModal(false);
-    setDeleteCohort(false);
-  };
-
-  const handleEdit = () => {
-    setEditShowModal(true);
-    setAnchorElUser(null);
-  };
-
-  const handleCloseEdit = () => {
-    setEditShowModal(false);
-    setIndicator(false);
-  };
-
-  const handleClickOpen = () => {
-    setShowModal(!showModal);
-    setIndicator(false);
-    setAnchorElUser(null);
-  };
-
-  const handleCloseEnroll = () => {
-    setEnrollShowModal(false);
-    setIndicator(false);
-  };
   const handleClickOpenEnroll = () => {
     setEnrollShowModal(!enrollShowModal);
     setIndicator(false);
-  };
-
-  const handleCloseUnenroll = () => {
-    setunenrollShowModal(false);
-    setIndicator(false);
-  };
-  const handleClickOpenUnenroll = () => {
-    setunenrollShowModal(!unenrollShowModal);
-    setIndicator(false);
-    setAnchorElUser(null);
   };
 
   const rolesList = user.data.user.rolesList;
@@ -133,74 +77,6 @@ function ClassCard() {
     ? (flag = true)
     : (flag = false);
 
-  // API CALL FOR DELETE CLASS
-  const deleteHandler = (id) => {
-    const notify = () => {
-      toast.success(" Deleted the class successfully", {
-        position: toast.POSITION.BOTTOM_RIGHT,
-        autoClose: 2500,
-      });
-    };
-    setShowModal(!showModal);
-    return axios({
-      method: METHODS.DELETE,
-      url: `${process.env.REACT_APP_MERAKI_URL}/classes/${id}`,
-      headers: {
-        accept: "application/json",
-        Authorization: user.data.token,
-        "delete-all": deleteCohort,
-      },
-    }).then(() => {
-      notify();
-      dispatch(classActions.deleteClass(id));
-    });
-  };
-
-  // API CALL FOR enroll class
-  const handleSubmit = (Id) => {
-    setLoading(true);
-    const notify = () => {
-      toast.success("You have been enrolled to class successfully", {
-        position: toast.POSITION.BOTTOM_RIGHT,
-        autoClose: 2500,
-      });
-    };
-    let getNotify = false;
-    setEnrollShowModal(!enrollShowModal);
-    const timer = setTimeout(() => {
-      getNotify = true;
-      setLoading(false);
-    }, 10000);
-    axios
-      .post(
-        // `${process.env.REACT_APP_MERAKI_URL}/classes/${Id}/register?register-all=${indicator}`,
-        `${process.env.REACT_APP_MERAKI_URL}/classes/${Id}/register`,
-        {},
-        {
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: user.data.token,
-            "register-to-all": indicator,
-          },
-        }
-      )
-      .then((res) => {
-        if (!getNotify) {
-          notify();
-          clearTimeout(timer);
-          setLoading(false);
-        }
-        dispatch(classActions.enrolledClass(Id));
-      })
-      .catch((res) => {
-        toast.error("Already enrolled in another batch", {
-          position: toast.POSITION.BOTTOM_RIGHT,
-          autoClose: 2500,
-          color: "red",
-        });
-        setLoading(false);
-      });
-  };
   // API CALL FOR DROP OUT
   const handleDropOut = (Id) => {
     setLoading(true);
@@ -320,47 +196,11 @@ function ClassCard() {
                 {((rolesList.length === 0 && item.enrolled) ||
                   (rolesList.length >= 1 &&
                     (item?.facilitator?.email === user.data.user.email ||
-                      flag))) && (
-                  <MoreVertIcon
-                    style={{ color: "#BDBDBD", cursor: "pointer" }}
-                    onClick={handleOpenUserMenu}
-                    sx={{ p: 0 }}
-                  />
-                )}
+                      flag))) && <EditClass item={item} />}
               </Typography>
               {/* dialog box for edit delete and merge class  */}
-              <Menu
-                sx={{ mt: "15px" }}
-                id="menu-appbar"
-                anchorEl={anchorElUser}
-                anchorOrigin={{
-                  vertical: "top",
-                  horizontal: "right",
-                }}
-                keepMounted
-                maxWidth="130px"
-                transformOrigin={{
-                  vertical: "top",
-                  horizontal: "right",
-                }}
-                PaperProps={{
-                  style: { width: "150px" },
-                }}
-                open={Boolean(anchorElUser)}
-                onClose={() => {
-                  setAnchorElUser(null);
-                }}
-              >
-                {(item?.facilitator.email === user.data.user.email || flag) && (
-                  <>
-                    <MenuItem
-                      onClick={() => handleEdit(item.id)}
-                      sx={{ width: 133, margin: "0px 10px" }}
-                    >
-                      <Typography textAlign="center">Edit</Typography>
-                    </MenuItem>
 
-                    {/* {ACBPathway?.code === "ACB" && !item?.merge_class && (
+              {/* {ACBPathway?.code === "ACB" && !item?.merge_class && (
                 <MergeClass
                   item={item}
                   itemID={item.id}
@@ -369,24 +209,6 @@ function ClassCard() {
                   setRefreshKey={setRefreshKey}
                 />
               )} */}
-                    <MenuItem
-                      onClick={() => handleClickOpen(item.id)}
-                      sx={{ width: 133, margin: "0px 10px", color: "#F44336" }}
-                    >
-                      <Typography textAlign="center">Delete</Typography>
-                    </MenuItem>
-                  </>
-                )}
-
-                {!rolesList.includes("volunteer") && item.enrolled && (
-                  <MenuItem
-                    onClick={() => handleClickOpenUnenroll(item.id)}
-                    sx={{ width: 120, margin: "0px 10px" }}
-                  >
-                    <Typography textAlign="center">Dropout</Typography>
-                  </MenuItem>
-                )}
-              </Menu>
 
               {/* it will show when two class merged */}
               {/* {ACBPathway?.code === "ACB" && item?.merge_class && (
